server/requirements.txt
Flask
numpy
pillow
